Geometry cache fixes

+++ test/check_scene.h
@@ -424,7 +424,7 @@ static void ufbx_check_line_curve(ufbx_scene *scene, ufbx_line_curve *line)
 {
 	for (size_t i = 0; i < line->point_indices.count; i++) {
 		int32_t ix = line->point_indices.data[i];
-		ufbxt_assert(ix >= 0 && ix < line->control_points.count);
+		ufbxt_assert(ix >= 0 && (uint32_t)ix < line->control_points.count);
 	}
 	for (size_t i = 0; i < line->segments.count; i++) {
 		ufbx_line_segment seg = line->segments.data[i];
